stm_miphy: Add provision for dedicated sata uports
Later hardware such as the ORLY2 has the SATA and PCIe chips integrated
differently to previous chips. In previous chips, the uport was present
only in the PCIe express controller, so even if you were programming the
miphy into sata mode you still talked to the miphy through the PCIe
uport. However, later chips have a uport present in both the PCIe
controller and the SATA controller, controlled by a mux. So if you are
programming the miphy for SATA you must talk via the SATA uport.

This patch introduces the concept of having a dedicated sata port into
the miphy code. If there is only a pcie uport, then it will be used for
both sata and pcie.

Signed-off-by: David McKay <david.mckay@st.com>