stm: b2044,b2078: make selectable eRouter for ISVE
The eRouter (if1) can be managed by either ISVE or fastpath.
We could enable the STM_ISVE_EROUTER only if the fastpath is disable
but this is a limit because we need the FP for the Eth connection.
In fact, STM_ISVE_EROUTER has no dependencies on BYOI.
Maybe, in the future STM_ISVE_EROUTER could only enable if FASTPATH_CMCI
is disabled.
With this patch the developers can manually enable the eROUTER to be managed
by ISVE but that does mean developers have to know what to do

Signed-off-by: Giuseppe Cavallaro <peppe.cavallaro@st.com>
Cc: Martin Habets <martin.habets@st.com>