stm_miphy: Introduce SATA GEN3 miphy settings
The Orly2 miphy and controller support SATA3 speeds, this patch adds the
necessary support into the miphy code in order to support this. Whether
you want to use SATA3 speeds is selected by setting the sata_gen
parameter.  The default is SATA1 on older revisions of the miphy and
SATA2 on the new versions.

The important point is that the controller must be able to support SATA3
speeds as well, thus the AHCI controller on the Orly2 can support SATA3,
but the legacy controller on the Orly1 cannot. Hence on Orly1 we are
limited to SATA2 speeds even though the miphy can support SATA3.

Signed-off-by: David McKay <david.mckay@st.com>