stm: armv7: pm: Fixed stm_suspend_on_eram_sz evaluation
This patch fixes the stm_suspend_on_eram_sz evaluation.
This is a variable in the suspend assembly code used to evaluate
the size of the 'stm_suspend_on_eram' assembly function.

But the variable was taking into account all the assembly file
instead of the only stm_suspend_on_eram function.

Now fixed in this manner less code is copied on ERAM.

Tested on: b2000-h415

Signed-off-by: Francesco Virlinzi <francesco.virlinzi@st.com>
Signed-off-by: Stuart Menefy <stuart.menefy@st.com>