stm: hom: h416: Move DDR in self-refresh using PCTL
This patch updates the H416 HoM table to move
the DDR in self-refresh using the Synopsys PCTL controller
instead of using the sysconf registers.

This is a hw bug on H415 (where it can't be done through the
sysconf registers) and the same code had been used for the H416
where, it seems, this bug doesn't exist.

In any case the sysconf support can be enabled just commenting
a define; this is done to simplify the changes in a secure system.

Signed-off-by: Francesco Virlinzi <francesco.virlinzi@st.com>
Signed-off-by: Stuart Menefy <stuart.menefy@st.com>