stm: b2020_h416: fix the on-board flash devices configuration
This patch fixes the support for Serial flash and NAND flash on b2020 orly-2
variant.

Fix the type of serial flash that is actually a Spansion SPI Flash.
Configure the NAND BCH controller.
Update the defconfig to enable the support for the SPI FSM and NAND BCH
driver. Further enable support for JFFS2 and UBIFS flash filesystems.

Signed-off-by: Nunzio Raciti <nunzio.raciti@st.com>
Signed-off-by: Carmelo Amoroso <carmelo.amoroso@st.com>
Signed-off-by: Stuart Menefy <stuart.menefy@st.com>