stm: b2000_h416: add support for NAND flash device
This patch adds the support for NAND flash on b2000 orly-2 by enabling
the NAND BCH controller.

Note that the b2000 board hasn't a nand device mounted on its pcb, this support
is provided just in case we have a daughter board  inserted on the CN43 connector.

Further defconfig is updated enabling support for SPI FSM and NAND
drivers, and JFFS2 filesystem.

Signed-off-by: Nunzio Raciti <nunzio.raciti@st.com>
Signed-off-by: Carmelo Amoroso <carmelo.amoroso@st.com>
Signed-off-by: Stuart Menefy <stuart.menefy@st.com>