stm_pcie: Add pipe interface to Miphy
There is a register block to control the PIPE interface between the PCIe
block and the Miphy itself. The PIPE interface is a standard for talking
between PCIe controllers and phys. This code allows the miphy code to
program the pipe registers if required to bring up a PCIe link.

Signed-off-by: David McKay <david.mckay@st.com>