stm: stig125: Added GPIO interrupt support
This patch adds the GPIO interrupt support on STiG125.

The PIO hardware consists of four banks of PIO blocks: West, SBC and
South use the PIO10 block which has an interrupt multiplexer with a
status register for each PIO block, while North is four standalone
PIO blocks without the extra multiplexer. However unlike previous
SoCs which used the PIO10 block and ORed together the interrupts
from the PIO blocks before they were visible to the CPU, on the
SASC1 all the interrupts are taken individually to a separate IRQ MUX
block, which reduces the 27 inputs down to 10 which are taken to the
CPUs.

To allow the existing drivers to be used unmodified, PIO blocks west,
SBC and south have all their interrupts ORed together in the
IRQ MUX, allowing the gpio-irq-mux to be used. The four interrupts from
the north block are mapped 1 to 1 to four A9 interrupts, so no
pio-irqmux is required.

Reference: [Bugzilla 26115]

Signed-off-by: Francesco Virlinzi <francesco.virlinzi@st.com>
Signed-off-by: Stuart Menefy <stuart.menefy@st.com>