b2020_h416: add MMC1
This patch adds the MMC1 that is actually connected to the eMMC
device on this kind of board.
MMC0 remains configured as removable becasue assigned to the SD/MMC slot.
MMC1 is now configured and marked as not removable because usef for the eMMC.
I have tested all on a revC and the Toshiba eMMC is working fine.

See Bugzilla Bug: 27042

Signed-off-by: Giuseppe Cavallaro <peppe.cavallaro@st.com>
Signed-off-by: Stuart Menefy <stuart.menefy@st.com>