mtd: nand: acknowledge nand_get_features is unused
The nand_get_features() function is currently unused and as such, generates a
compile time warning.  However, the function is useful during development and
debugging, and may be used in its own right in the future.

This patch adds the __maybe_unused decoration to acknowledge the fact that the
function is unused and to suppress the compiler warning.

Signed-off-by: Angus Clark <angus.clark@st.com>