stm_nand_bch: add support for devices that use 4 address cycles
Previously the stm-nand-bch driver was hard-coded to use 5 address cycles for
Read and Write operations, and 2 address cycles for Erase operations.  This
patch adds support for NAND devices that use 4 and 2 address cycles for
Read/Write and Erase operations respectively.

Note, this patch has not been tested on any real NAND device; I have failed to
find a BCH-capable ST reference board that is populated with a '4-address-cycle'
NAND device.

Signed-off-by: Angus Clark <angus.clark@st.com>