stm_nand_bch: relax tests on permitted OOB operations
The BCH driver does not allow use of the OOB for anything other than ECC data.
As such, OOB operations using modes 'MTD_OPS_AUTO_OOB' and 'MTD_OPS_PLACE_OOB'
are not permitted.

Previously, the driver just tested the 'mode' variable, and returned -ENOTSUPP
if an incompatible mode was specified.  However, it turns out that some clients
(eg YAFFS2) set the OOB mode, even if OOB operations are not going to be
performed.

This patch updates the code to test the mode setting only if we are actually
going to perform a OOB operation, as indicated by a non-NULL 'oobbuf' being
passed in.

Signed-off-by: Angus Clark <angus.clark@st.com>