mtd: nand: workaround for mismatch on MX30LF1G08AM READID discovery
During device initialisation, the READID signature is probed twice to ensure
random data on the bus cannot be misinterpreted as a valid NAND device.

However, on the second READID probe, the Macronix MX30LF1G08AM device seems to
continue from where the first probe finished, rather than restarting from the
first byte of the READID signature.  As a result the probe fails, reporting a
mismatch on the READID data.

This patch introduces an additional device RESET prior to the second READID
probe.  This avoids the READID mismatch on the MX30LF1G08AM and should be benign
on other devices.

This issue has been reported to Macronix. Awaiting their response.

Signed-off-by: Angus Clark <angus.clark@st.com>