stm: stih416: improved the power consumption in Standby
This patch improves the power consumtion during Standby on H416
by disabling additional clocks.

This result is possible only by changing which ERAM the CPU uses:
an internal undocumented ERAM is used which is "closest" to the
CPU (the lastest 2K of this ERAM are used).

Tested on: b2000-h416

Signed-off-by: Francesco Virlinzi <francesco.virlinzi@st.com>
Signed-off-by: Stuart Menefy <stuart.menefy@st.com>