stm_sata: Enable SSC (spread spectrum clocking)
As all the ST SATA IPs have SSC_en wired up to phy_ctrl[10], it is
recommended by validation to enable SSC by default.

SSC is the technique of modulating the operating frequency of a signal
slightly to spread its radiated emissions over a range of frequencies.
This reduction in the maximum emission for a given frequency helps meet
radiated emission requirements.

Signed-off-by: Harsh Gupta <harsh.gupta@st.com>
Signed-off-by: Srinivas Kandagatla <srinivas.kandagatla@st.com>
Signed-off-by: Stuart Menefy <stuart.menefy@st.com>