stm: stih416: consolidate UARTs OE pad settings
Mostly 'not_OE' PIO pin is not used even if configured, and in some
case it's not configured at all.

This patch align the UART pads in order to configure the OE pin always
when available, to be consistent with SoC architecture, and extend the
ASC board configuration with an flag to specify if the 'not_OE' pin
has to be claimed or not (default is not claim).  This will provide
more flexibility in managing shared PIO pins.

On orly-2, the 'not_OE' pin of UART_2 (PIO17_3) is shared with MMC0_LED.
If UART_2 is configured (on b2105 is the kernel console), it was claiming also
the 'not_OE' PIO making the sdhci driver to fail and reset. As in this use case
the 'not_OE' is not functionally required to use UART_2, by ignoring it we can
have both console and MMC0 working.

Signed-off-by: Carmelo Amoroso <carmelo.amoroso@st.com>
Signed-off-by: Stuart Menefy <stuart.menefy@st.com>