stm:snd Remove unused audio subsystem module parameters
This patch removes the module parameters 'index' and 'id' from the ALSA
modules for FLi7610, STiH415, STiH416, and STx7108 SoCs. The parameters
are not used anywhere inside these modules and are most likely a remnant
of an initial module copied from core.c.

The module parameters 'index' and 'id' have also been removed from core.c
as they are not used.

Signed-off-by: John Boddie <john.boddie@st.com>