stm:snd Simplification of ALSA SoC glue drivers
This patch simplifies the ALSA SoC glue drivers by using the STM PAD API
to claim and configure the specified SYSCONF fields.

The simplification has also allowed power management support to be added
to glue drivers missing it. The power management support is based on the
STiH416 glue driver power management support and merely ensures the SYSCONF
registers are re-initialised to the correct values.

Signed-off-by: John Boddie <john.boddie@st.com>