stm:fdma Sanity check residue value
Whilst developing the TELSS ALSA driver, in certain circumstances when
calculating the residue of the current DMA transfer, we read an invalid
value for the current node's residue from the NODE_COUNT_REG. i.e. the
value read is greater than the total transfer length.

This patch modifies the stm_fdma_get_residue() function so that when we
search the transfer descriptor for the current transfer node, on finding
it, we check to see if the NODE_COUNT_REG value is larger than the actual
length of the node. If it is we use the node length as the node residue.

This will have the effect that we will return an inaccurate transfer
residue. The residue will be overstated by as much as size of the current
node, but on the next call to get the residue function the value returned
should be more accurate. This is more preferable to understating the
transfer residue.

This patch also sets the correct return type for the stm_fdma_get_residue()
function. It is a u32, not an int.

Signed-off-by: John Boddie <john.boddie@st.com>