stm:fdma More thorough TELSS parameter checking
This patch fixes a simple off by one error where we check the TELSS handset
count is between 1 and 9 instead of between 1 and 10 and adds more thorough
parameter checking for the frame count and frame size parameters.

An additional parameter has also been added to the TELSS configuration to
allow endian swap to be enabled or dissabled.

Signed-off-by: John Boddie <john.boddie@st.com>