stm:fdma Add period stride parameter to TELSS transfer API
This patch adds an additional parameter to the dma_telss_prep_dma_cyclic()
API call named period_stride. The purpose of this parameter is to specify
the stride between periods and is used when calculating the buffer offset
for each FDMA node in a given TELSS transfer. Adding the period_stride
parameter also allows for the start of each period on a given alignment.

Signed-off-by: John Boddie <john.boddie@st.com>