stm:fdma Additional API to get the current TELSS period being processed
This patch provides a new function to return the current period of an
executing transfer for a specified TELSS channel.

The ALSA sound system periodically updates it's internal buffer pointers
based on the progress of a DMA transfer. Normally the progress is
determined by calculating the DMA transfer residue. For a TELSS channel,
the residue is not reliable, as the TELSS channel may zero stuff it's
output. As such we can only reliably determine what DMA transfer node
(or period) is being transferred. This new API returns the currently
being processed period.

Signed-off-by: John Boddie <john.boddie@st.com>