stm:snd:uniperif Calculate FDMA trigger limit based on channel count
This patch replaces the arbitrary method of calculating the FDMA trigger
limit with a new method based on the number of channels and uniperipheral
FIFO size.

Essentially, to ensure that sync over multiple uniperipheral players (with
different numbers of output channels) can be maintained (particularly when
the players are stopped), the FDMA trigger limit is calculated to ensure
only a fixed number of frames is ever in the FIFO at any time. Previously
the FDMA trigger limit was calculated to allow the maximum number of frames
in the FIFO at any time.

Signed-off-by: John Boddie <john.boddie@st.com>