arm/PCI: Allow multiple calls to pci_common_init()
pci_common_init() can only be called once at the moment, because it
calls pcibios_init_hw() which starts numbering PCI busses from zero.
Normally this isn't a problem as you can instantiate any number of
controllers using this call using the nr_controllers field. However, on
some hardware that has PCI or PCIe controllers that are *totally*
different, it is easier to allow multiple calls to pci_common_init().

All that has to be done is make the starting bus a static variable,
which will ensure that we start with a fresh bus number at each call of
pcibios_init_hw().

Signed-off-by: David McKay <david.mckay@st.com>