stm:miphy: Add 10 bit symbol setting for PCIE
For the SASC1 PCIe cell it is necessary to set the symbol size to 10
bits rather than the default of 20. This is done through the PIPE
interface connected to the approprate PCIe controller.

For dual lane phys, such as on the SASC1, setting the pipe registers
should only be done once, as the settigns are broadcast to the other
lane anyway. Thus there is flag set to ensure that the pipe settings are
done only for the first of the controllers to be configured.

Signed-off-by: David McKay <david.mckay@st.com>