stm:stig125: Enable PCIe on SASC1 on b2044 board
This patch adds support for the PCI Express cells on the SASC1. The
B2044 has two PCI slots, connected to controllers 1 and 2. The other
potential PCIe slot is dedicated to SATA instead.

Some more work to configure the clocks and phy multiplexor will be
required if PCIe controller 0 is used on another board, but this should
be straightforward.

Signed-off-by: David McKay <david.mckay@st.com>