stm:snd:stih41x Provide Kconfig options to select B2011A audio input/output
This patch is associated with Bugzilla 16138 and provides a Kconfig option
to enable the use of B2011A with the B2000 board. The B2011A board uses
shared PIOs with both the I2C and ASC/UART.

The Kconfig allows selection of either audio input or output mode and will
automatically configure the PIOs for the specified audio mode. It will also
set an alternate console if the default cannot be used.

Also in this patch is a minor modification to allow PIOs with 8 alternate
functions to be used on STiH415.

Signed-off-by: John Boddie <john.boddie@st.com>
[stuart.menefy@st.com: Moved Kconfig changes to Kconfig.b2000]
Signed-off-by: Stuart Menefy <stuart.menefy@st.com>