stm/of/StiH415: Add device tree support.
This patch add device tree support to StiH415, Now all the SoC support
core is conditional compiled with CONFIG_OF. The code which is not under
CONFIG_OF will be remove later.
From device tree side: The files are splitted as
 - SOC support nodes.
 - Pad config nodes.
 - Ethernet configuration nodes.

By default most of the devices have `status = "disabled";` property,
which means the device will not be instatianted by the of_platform code,
The BSP level dts file should make this property as okay to get the
device enabled.

Signed-off-by: Srinivas Kandagatla <srinivas.kandagatla@st.com>