stm/of/B2044: Convert BSP to device trees.
This patch converts b2044 bsp support to device trees. This file
ONLY exists to easy the transission of sdk2 to device trees. Doing this
way will allow both old style and device trees based devices to
co-exist.

Also with this approch sdk2 gets same kernel config without device trees
and also device tree support.

Users should be carefull of what they add in this file as part of the
BSP support goes via device trees in this approch, and make sure that
they are not duplicating in both places.

Signed-off-by: Srinivas Kandagatla <srinivas.kandagatla@st.com>