stm: nand: simplify method for determining MTD slave offsets
The stm-nand-flex and stm-nand-afm drivers include an option for supporting
boot-mode ECC, as used by the STM Hamming NAND controller.  Since boot-mode ECC
is not generally suitable for other purposes (e.g. hosting NAND filesystems),
boot-mode ECC is confined to a single, named, partition, with all other
partitions using the driver's "native" ECC scheme (i.e. 'linux' or 'afm', for
stm-nand-flex and stm-nand-afm respectively).

To support this mode of operation, the driver must be able to determine the
bounds over which boot-mode ECC should be applied.  Previously this was achieved
by adding a function for returning the named partition's mtd_info structure, and
then using a duplicated, local definition of mtdparts's internal data structure
to get access to the slave's offset.  At best, this should be regarded as a
hack.

This patch updates the function for retrieving the slave partition such that it
also returns the slave's offset.  This keeps the API cleaner, and avoids the
need to maintain local copies of mtdpart data structures!