stm: pm: make suspend work with an identity map
Due to CONFIG_VMSLIT_1G in b2020 platform kernel memory (virtually) sits
at 0x40000000 but also the physical memory is at 0x40000000.

During PM operation the MMU is disabled and after that it's enabled...
but due to the fact the CPU is moving from "physical" 0x40... to
"virtual" 0x40... this can create aliasing (and data corruption)
as the different mappings have different attributesr, so we have to be
very careful performing these transitions, flushing caches and TLB at
appropriate times.

This patch switches the code to use the generic C functions when
switching to and from identity mappings, which get this right.

Tested on: b2020-h416 and b2000-h416

Reference: [Bugzilla 26535]

Signed-off-by: Francesco Virlinzi <francesco.virlinzi@st.com>