ata: ahci_stm: Add MiPHY number parameter to platform data
Currently the code assumes that the MiPHY number will be the same as
the AHCI platform device ID. While that may be true currently, there is
no guarantee that it will be true in the future, so add a parameter
allowing it to be specified explicitly.

This also matches the sata_stm code.

Signed-off-by: Stuart Menefy <stuart.menefy@st.com>