ata: ahch_stm: Move functionality into ahci platform callbacks
Prior to this patch the stm specific functionality was split
between the ahci platform device callbacks and the ahci_stm platform
device probe and power management functions.

However, this split makes it harder to determine which device (the
ahci or the ahci_stm device) is managing a particular resource. This
is in contrast to the other systems which use the ahci platform
device, where all the code is in the callbacks.

So shift all the code into functions which are called directly by the
ahci platform device, leaving just the initialisation of the ahci
platform device in the ahci_stm platfrom device's probe function.

As part of this change use of the pm_clk_... functions have been
replaced by direct clk_... calls, as they don't do what we want
when power management has been disabled.

This should also make it easier to rework this code for device trees in
the future.

Signed-off-by: Stuart Menefy <stuart.menefy@st.com>