stm: b2044: Remove Tx polarity invert on SATA port 1
The old MiPHY A-40x code only supported a single MiPHY, and returned
an error code if an attempt was made to initialise any device other
than number zero. On the STiG125 which has two MiPHYs, the initialisation
attempt on the second device failed, but then the error was being ignored
by the AHCI wrapper device, which continued to bring up the AHCI
device driver. However, as the MiPHY was being correctly initialised
by the hardware for SATA operations, it appeared to work correctly.

A result of this was that any polarity configuration being passed into
the MiPHY driver had no effect for the second MiPHY.

However, commit 0710324be7f0600fb85ecea1c8ca7ccd8d28a633
    stm_miphy: Allow multi-instance miphy code
changed the MiPHY code so that the second MiPHY is now correctly
supported.

As a result polarity inversion fields were now being acted on, and
it was found that specifying Tx polarity inversion for the second
port is erroneous, and needs to be removed.

Signed-off-by: Stuart Menefy <stuart.menefy@st.com>