stm_nand_flex: increase tWB delay for RBn polling
The "Ready/Not-Busy" (RBn) signal is used to indicate when the NAND device has
completed certain operations.  After receiving a command, a small delay is
permissible before the NAND device deasserts the RBn signal.  This delay is
usually referred to as 'tWB' in the datahseets, and is typically 100ns.

To avoid a race condition, the software drivers must apply a suitable delay
before attempting to sample the RBn status.  Previously, it was assumed a delay
of 100ns was sufficient.  However, following the "Timing Spec" updates (which
allow us to drive the NAND Controller closer to the specification limits) it
seems we have revealed another issue with the Hamming NAND Controller.

Under certain conditions, it was found that polling the RBn status register
would return immediately, and we would end up with errors indicative of the
operation not having completed.  Further tests found that the correct behaviour
was restored when increasing the delay before polling the RBn register.

The issue has been reported to the IP and Validation Teams.  It is suspected
that the NAND Controller requires a number of clock cycles before the RBn
register is updated, which in effect introduces a lag in what is reported.

While waiting for official confirmation of the issue, this patch increases the
delay from 100ns to 200ns.  This value is based on empirical tests, plus a
safety margin.  This should have little impact on overall performance, since
most operations that involve polling the RBn status require anywhere between 25
to 1000us.