stm_nand_bch: allow some non-OOB operations with mtd_{read,write}_oob
Certain MTD clients (e.g. YAFFS2 with 'inband-tags' enabled) use the
mtd_read_oob() and mtd_write_oob() calls for performing operations that do not
actually involve OOB data.  (In the case of YAFFS2, this is probably down to the
way in which 'inband-tags' support was bolted on to the existing code.)

This patch adds a test for non-OOB requests, and then passes control to the
conventional mtd_{read,write} calls as appropriate.  Requests for 'raw' data
(i.e. mode = MTD_OPS_RAW) are still handled by the mtd_{read,write}_oob calls,
which make use of Hamming-FLEX mode operations to fulfill the requests.