stm:telss Addition of ALSA uniperipheral TDM driver [v2]
This patch contains the initial version of the uniperipheral TDM driver.
All platform configuration data must be specified to the driver - due to
the multiple TELSS hardware configurations, the driver will not attempt to
work things out for itself!

On probe, the driver will start the uniperipheral, but with interrupts and
dma disabled. This will allow the uniperipheral clock to propogate down to
the TELSS hardware. The clock will only be stopped upon removal of the
driver. The dma will be started and stopped as required, i.e. only when
there is a call to process.

The driver provides a pcm interface per handset. So the uniperipheral TDM
player will provide a pcm player interface per handset, and the
uniperipheral TDM reader will provide a pcm reader interface per handset.
These interfaces should be the same as any other pcm interface. The table
below shows what ALSA formats correspond to what TELSS call types:
	C/NB:	S8, U8
	L/NB:	S16_LE
	C/WB:	S16_LE
	L/WB:	S32_LE

The call type is selected automatically depending on the ALSA format, but
the user application will still have to inform the TELSS hardware what
format to expect.

No power management is provided. Future versions may reduce the PCLK but
nothing more, as the DECT/Lantiq/ZSI modules must always be supplied with
a clock.

The uniperipheral driver performs no control of the TELSS hardware or the
TELSS top glue. You must provide your own driver for this that will also
configure things such as the PIO etc.

This driver was initially designed and implemented around assumptions for
the Lantiq DuSLIC - these may not be valid for other devices and may end
up requiring driver modifications.

Signed-off-by: John Boddie <john.boddie@st.com>