stm:of:fdma Increase IMEM size as define in device tree
On STiG125 there is an issue loading the FDMA firmware where the error
indicates that the firmware segment is too large to load.

The FDMA has 8192 words of memory for 24Kb total. Each word is actually
24-bits not 32-bits. When we convert the FDMA firmware into an ELF file,
the instructions are padded from 24-bits to 32-bits. Consequently, the
size of the .data segment will grow by 25% which may result in it being
larger than 24Kb (when the section is written to IMEM the padding is
skipped!).

This patch increases the FDMA IMEM size from 24Kb to a padded 32Kb.

Signed-off-by: John Boddie <john.boddie@st.com>