stm:of:stig125 Fix SPI 12/13 configuration
This patch corrects the definitions of the SPI 12/13 configuration for
STiG125. SPI13 is used for TELSS. As part of this change the pad config
for SPI13 has been changed from all pins being bi-directional to all pins
being either in or out.

The patch also takes the opportunity to rename the SPI device tree nodes
to something more descriptive, e.g. telss_spi0 instead of spi13, and to add
an alias for the TELSS SPI.

Signed-off-by: John Boddie <john.boddie@st.com>
Acked-By: Srinivas Kandagatla <srinivas.kandagatla@st.com>