stmmac: dirty hack to fix dual gmac when use OF
In case of multiple instance of the stmmac it is necessary to allocate
own platform structure otherwise there is the risk that the private
structure will always point to the latest pdev->dev.platform_data.
This means that the eth0 could have the platform used for eth1.
Also to fix the callback pointers, as soon as, in the stmmac_pltfr_probe
function the OF is invoked and all the DT fields are setup, the
pdev->dev.platform_data has to point to the new platform.
This approach is not the best one but it allows us to avoid to change
the driver APIs and have a fast and dirty fix for configuring dual
MAC on ST boards.

Signed-off-by: Giuseppe Cavallaro <peppe.cavallaro@st.com>
Reviewed-by: Srinivas Kandagatla <srinivas.kandagatla@st.com>