stm/of/hcd-stm: remove setting up devname and pdev->id.
Setting pdev->id or pdev name dynamically is illegal, this patch cleans
up the way the device tree support is implemented.
Now the pdev->id and platform data is set in a local variables
rather than in pdev.

Also this patch adds aux_data entries in respective boards to accomdate
this change.

Signed-off-by: Srinivas Kandagatla <srinivas.kandagatla@st.com>