hwmon: stm-pwm: Change how configuration information is passed in
The current PWM driver requires a single flag for each channel,
whether it should be enabled or not. This is passed in as an array
of ints in platform data or from the device tree. This makes it
ugly to add additional parameters.

So switch to a model of passing in a per-channel structure. In
platform data is a simple struct, in device tree it is a sub-node,
which by default is disabled, and needs to be enabled in the normal
way to enable the channel, for example:

pwm1: pwm@fe510000 {
	status = "okay";
	pwm-channel@3 {
		status = "okay";
	}
};

All the new structure and field names have "pwm" in them, making it
easier to support the capture channels if this is ever required in the
future.

Signed-off-by: Stuart Menefy <stuart.menefy@st.com>