stm: sysconf:  fix sysconf debugfs support
This patch fixes the debugFS support for the sysconfig that generates a
panic while doing a simple: cat /sys/kernel/debug/sysconf. This was due
to the null plat pointer in the sysconf_block structure that is now set
in this patch in the probe(). In the non device tree case this is
assigned in the sysconf_early_init(). It cannot be done that way in
the device tree case as the list of platform devices is not available,
so it is assigned in the probe.

This support remains very helpful especially when bring-up SoC, boards
and testing device configurations.

Signed-off-by: Giuseppe Cavallaro <peppe.cavallaro@st.com>