stm: stig125: fastpath: add a config optional data to configure fp
Extend the API to configure the FastPath by adding a config data as
argument (that can be NULL when not needed) to pass platform
configuration. Currently it is supporting only the phy address, as in the next
b2112 board (alicante cut 0) to prevent potential contention on the MDIO line
the address the CPE PHY (U30) address is changed from 0 to 4 (it needs to move
RE11 to RE18)

As the FastPath configure all its interface all at once (docsis, isis,
fpgige), the config data consists of an array of if configuration, one
for each interface.

Signed-off-by: Carmelo Amoroso <carmelo.amoroso@st.com>
Signed-off-by: Giuseppe Cavallaro <peppe.cavallaro@st.com>