stm: stig125: change TX an RX clock retiming for FastPath BYOI
On B2112 (Alicante Cut 0) board, TX & RX lock delay has to be modified
to address packets lost issue.
This latest modifications have resolved the packate lost issue on
several boards.

The following hw modifications are required:

- C326 removed
- RE17 moved to RE10
- RE16 moved to RE9

Although this is not the best approach it is helping to fast fix the
problem on this kind of HW. Next time, a pltf field, maybe, could be
used for changing at run-time the retiming w/o adding any Koption
in the pad.

Reported-by: Jeremy Gasper <jeremy.gasper@st.com>
Signed-off-by: Carmelo Amoroso <carmelo.amoroso@st.com>
Signed-off-by: Giuseppe Cavallaro <peppe.cavallaro@st.com>