stm: hom: Add self-reset for debug/testing
This patch adds a self-reset to debug/test HoM.
When enabled as the latest operation in the poke_table is
an hard-reset forse an HoM resume.

This patch can be used in environment where SBC/LPM ins't integrated
moreover in this manner HoM debug/testing doesn't need direct access
to the board.

Signed-off-by: Francesco Virlinzi <francesco.virlinzi@st.com>