stm: b2105: add board support
This patch adds the initial support for the b2105 board based on DT.

It almost includes all the devices like MMC, NAND GMAC0 and adds the
defconfig.
GMAC1 is not enabled by default. It will be worth enabling it as soon
as Alicante will be added too (so the two boards will be able to
communicate via NIM).

V2: fixed memory settings

Contributed-by: Carmelo Amoroso <carmelo.amoroso@st.com>
Contributed-by: Nunzio Raciti <nunzio.raciti@st.com>
Signed-off-by: Giuseppe Cavallaro <peppe.cavallaro@st.com>