stm: stih416-b2000: turn-on GMAC1 on CN23
This patch enables the GMAC1 to work in GMII mode
and fixes a bug in the bus_id that has been changed from
0 (used by GMAC0) to 1.

IMO this can make sense for several reasons; for example
the B2032 (modified for GMII) is mostly used in this board,
the CN22 can be used for MMC devices so the GMAC1 is mandatory.
Dual GMAC can be supported simultaneously.

See Bugzilla: bug 31273

Signed-off-by: Giuseppe Cavallaro <peppe.cavallaro@st.com>