stm:telss Do not define TELSS card type by default
This patch addresses Bugzilla 30855. When TELSS is not configured, if we
do an aplay -l on SDK2 we see devices for sound card 0, 2, 3, etc, but not
card 1 (which is the TELSS card).

This patch will ensure that when TELSS is not configured that the cards
listed by aplay -l are contiguous.

Signed-off-by: John Boddie <john.boddie@st.com>