stm: i2c: fsm-idle state fixed to avoid glitch on SDA line
This patch fixes the fsm-IDLE state to avoid a glitch on SDA line.

With the commit b8ca463b363a7d9d8bb9c90177f84551795946cd the
SDA pull-up was moved in the on process context but in the idle
state the code is still resetting the 'STOPG' bit in the SSC_I2C control
register; due to that nobody is still pushing SDA line high.

This patch postpone the write operation on SSC_I2C to be done
after the gpio_sda is already updated.

Reference: [Bugzilla 31661]

Signed-off-by: Francesco Virlinzi <francesco.virlinzi@st.com>