stm: rtc: allowed more than one RTC-LPC wokenup
With this patch the rtc-lpc can woken-up more than once.

The problem was that due to lpc-hw bugs on resume it was agreed a
complete lpc reset.

But if the rtc-lpc is a wokenup device the kernel waits for ever an
rtc_update_irq which will never be called due to the lpc reset,
and without the rtc_update_irq the kernel will reject other rtc-lpc
programmation.

With this patch when the rtc-lpc is a woken-up device on resume the
driver calls the rtc_update_irq to notify a
virtual-interrupt which will allow further rtc-lpc programmation.

Tested-on: b2020-h415
Reference: [Bugzilla 31575]

Signed-off-by: Francesco Virlinzi <francesco.virlinzi@st.com>